refactor(app): split economic condensed and load-list row helpers
Unblock the deferred wave-4 secondary lib splits now that product issues
#3985/#4002 are past implementation—keep overlay and dialog files lean
under the main-line budget (Refs #4013).
