+++ SPEC/program/repo-lint.md
@@ -301,7 +301,7 @@ Large `app/test/**` widget/screen suites that are split into numbered fragments 
 
 **Approved `app/test/support/` harness modules (Refs #3847, #3952, #4013):** Shared widget-test scaffolding MUST live under `app/test/support/` — not re-declared in individual test files. Canonical modules include `app_shell_harness.dart` (`buildAppShell`, `pumpAppShell`), `min_viewport_harness.dart`, `diplomacy_panel_test_support.dart` (`buildDiplomacyPanel`, `buildDiplomacyPanelShell`, `buildCivilianPanel`, `relativePowerSpanColors`, `pumpDiplomacyPanelBuilt`, `DiplomacyPanelBusDialogHost`, `CivilianPanelBusDialogHost`), `new_game_leader_selection_dialog_test_support.dart` (`pumpNewGameLeaderSelectionDialog`, `ensureTapNewGameLeaderSelectionStart`, `ensureTapNewGameLeaderSelectionCancel`), `combat_ui_specs_test_support.dart` (`combatUiSpecsFrame`, `combatUiSpecsDarkFrame`), `move_dialogs_specs_test_support.dart` (`moveDialogsSpecsFrameWithOpener`), `dialogue_overlays_specs_test_support.dart` (`pumpDialogueOverlaysUntilSettled`, `dialogueOverlaysLibraryUnitSource`, `wrapGameStartIntroOverlay`), `screen_spec_acceptance_test_support.dart` (`buildScreenSpecMainMenu`, `woodPanelButtonFinderFor`, `findGradientSurfaceFor`, `pumpScreenSpecMainMenuAtSize`), `game_screen_test_support.dart` (`buildGameScreenShellOverrides`, `buildGameScreenHost`, `buildGameScreenShellToGameFlow`), `production_panel_test_support.dart` (`buildProductionPanel`, `productionPanelTestFullPlayer`), `widgetbook_test_harness.dart`, `province_overlay_test_harness.dart`, `province_overlay_dark_token_scenarios.dart` (`regionMapWithCellVisibility`, `gameWithMilitaryDarkTokenUnit`, `gameWithFleetAndPendingNavalMove`, `sparseOverlayGame`, `gameWithNoFleets`, `regionMapWithLandCells`, `gameWithGrainTilesForOverlay`, `gameWithCivilianUnitsForOverlay`, `omniscientPlayerViewForTiles`, `overlayDarkTokenTileKey`), `editorial_monocle_dark_token_assertions.dart` (`expectMutedSingleSource`, `expectMutedObfuscated`, `expectEditorialMonocleDarkChrome`), `game_map_area_state_logic_test_support.dart` (`expectedBuildImprovementEnabledFromPipeline`, `expectPortFleetMarkersMatchTownPortDrawables`), `military_units_panel_test_support.dart` (`buildMilitaryPanel`, `expandFirstArmyExpansion`, `expandAllArmyExpansions`, `pumpArmySplitHarness`, `ArmySplitTestHarness`, `buildMilitaryProvinceTileLookupGame`, `buildMilitarySeaFleetDisplayGame`, `buildMilitaryArmyAtLisbonDisplayGame`, `buildMilitaryHomeArmyAtCapitalGame`, `buildMilitaryAdjacentOwProvincesTopology`, `buildMilitaryTwoFieldArmiesAtProvinceGame`, `buildMilitaryFieldArmyWithAdjacentOwnedGame`, `buildMilitaryCrossRegionOwnedMoveGame`, `buildMilitaryInvasionAdjacentHostileGame`), `naval_units_panel_test_support.dart` (`buildNavalPanel`, `wireNavalSplitForWidgetTest`, `wireNavalTransferForWidgetTest`, `buildNavalPanelHomeFleetOnlyGame`, `buildNavalPanelNamedSeaZoneGame`, `buildNavalPanelShipLabelGame`, `buildNavalPanelOwFleetsGame`, `buildNavalPanelCapitalHomeAndPeersGame`, `buildNavalPanelCapitalMergePortFleetsGame`, `buildNavalPanelSingleSeaFleetGame`, `buildNavalPanelEmptyHumanGame`, `buildNavalPanelBeachheadMissionGame`, `buildNavalCapitalAdjacentSeaTopology`, `buildNavalTwoSeaZonesTopology`), `trade_screen_test_support.dart`, `golden_capture_harness.dart`, `yarn_test_fixtures.dart`, and the `panel_fixtures/` barrel (`buildDiplomacyPanelGameWithNoDiscoveredFactions`, `buildDiplomacyPanelGameWithTribeDiscoveredByVisibility`). AST gates `check_app_test_no_duplicate_scaffolding.dart`, `check_app_test_no_duplicate_diplomacy_host.dart`, `check_app_test_no_duplicate_golden_host.dart`, `check_app_test_no_duplicate_yarn_bundle.dart`, `check_app_test_no_duplicate_dark_token_asserts.dart`, and the Widgetbook `_useCase` pattern in the scaffolding gate enforce this.
 
-**App test file size (Refs #4013):** `repo.app_test_file_size` caps every `app/test/**/*.dart` file at **1000 physical lines**. Files over the cap must appear on the shrink-only allowlist in `tool/check_app_test_file_size.dart`; adding a new oversized file requires removing another allowlist entry or shrinking first. Stale allowlist entries (missing path or file now ≤ cap) fail CI. After the wave-4 `ct_region_map_widget` split, the allowlist is **empty**. Wave-4 densify also splits `game_screen_narrow`, `production_panel`, and `diplomacy_panel` into `*_partN_test.dart` and consolidates duplicated GameScreen shell overrides into `game_screen_test_support.dart`.
+**App test file size (Refs #4013):** `repo.app_test_file_size` caps every `app/test/**/*.dart` file at **1000 physical lines**. Files over the cap must appear on the shrink-only allowlist in `tool/check_app_test_file_size.dart`; adding a new oversized file requires removing another allowlist entry or shrinking first. Stale allowlist entries (missing path or file now ≤ cap) fail CI. After the wave-4 `ct_region_map_widget` split, the allowlist is **empty**. Wave-4 densify also splits `game_screen_narrow`, `production_panel`, and `diplomacy_panel` into `*_partN_test.dart` and consolidates duplicated GameScreen shell overrides into `game_screen_test_support.dart`. Secondary lib splits (Refs #4013): Economic Extraction/Available condensed helpers in `province_sea_zone_detail_overlay_economic_condensed.dart`; load-list row/meta helpers in `load_game_list_dialog_row.dart`.
 
 **Feature-screen top-bar helper (Refs #3952):** Diplomacy / Production / Technology / Trade screens build Map-back + 18×18 icon + title chrome via `app/lib/widgets/game_feature_screen_top_bar.dart` (`GameFeatureScreenTopBar.build`); screen-specific `topBarKey` / title / icon asset constants remain on each screen.
